// src/domain/place/field/field-type.enum.ts
export enum FieldType {
    FOOTBALL_INDOOR = "FOOTBALL_INDOOR",
    FOOTBALL_OUTDOOR = "FOOTBALL_OUTDOOR",
    BASKETBALL_INDOOR = "BASKETBALL_INDOOR",
    BASKETBALL_OUTDOOR = "BASKETBALL_OUTDOOR",
    TENNIS_INDOOR = "TENNIS_INDOOR",
    TENNIS_OUTDOOR = "TENNIS_OUTDOOR",
    RUGBY_INDOOR = "RUGBY_INDOOR",
    RUGBY_OUTDOOR = "RUGBY_OUTDOOR",
    SWIMMING_POOL_INDOOR = "SWIMMING_POOL_INDOOR",
    SWIMMING_POOL_OUTDOOR = "SWIMMING_POOL_OUTDOOR",
    NONE = ""
}